Dodanie stron po próbie wysłania maila
