+++ main.py
@@ -1,7 +1,7 @@
 from flask import Flask, render_template, request,jsonify
 import os
 
-app = Flask(__name__,static_url_path="/static/logo/")
+app = Flask(__name__, static_folder='styles')
 
 
 @app.route('/')
@@ -19,14 +19,10 @@ def sendData():
     mail = request.form.get('mail')
     message = request.form.get('message')
     if send.checkMail(mail) == True:
-        send.sendMail(mail,message)
-        #TODO zrobić szablon strony aby później można było dziedziczyć z niego i dodawać tylko do niego poszczególne elementy
-        #TODO jsonify("Email został wysłany") należy zastąpić  render_template('email_sent.html')
-        #TODO należy tam umieścić całą stronę internetową która ma zostać wyświetlona w przypadku gdy wiadomość  została pomyślnie wysłana
-        return jsonify("Email został wysłany")
+        send.sendMail(mail, message)
+        return render_template('email_sent.html')
     else:
-        #TODO zrobić podonie z jsonify("Email nie został wysłany") jak wyżej
-        return jsonify("Problemy z wysłaniem emaila")
+        return render_template('email_error.html')
 
 
 @app.route('/about_us/')
