--- scriptJS/main.js
@@ -1,3 +1,0 @@
-import { changeNavbar } from "./navbar.js";
-
-changeNavbar();
