+++ static/scriptJS/main.js
@@ -1,0 +1,3 @@
+import { changeNavbar } from "./navbar.js";
+
+changeNavbar();
