+++ static/styles/_hamburger.scss
@@ -1,0 +1,43 @@
+.hamburger {
+  margin-right: 1rem;
+  height: 60%;
+  width: 60px;
+  padding: 10px;
+  border-radius: 10px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  transition: 0.5s all ease-in-out;
+  justify-content: space-around;
+
+  .bar {
+    display: flex;
+    height: 5px;
+    border-radius: 10px;
+    background-color: var(--secondary-color);
+    background-color: var(--secondary-color);
+    width: 100%;
+    transition: 0.5s all ease-in-out;
+  }
+}
+.hamburger-open {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  transition: 0.5 all ease-in-out;
+}
+.hamburger-open .bar:nth-child(1) {
+  transform: rotate(225deg) translateY(-7px);
+}
+.hamburger-open .bar:nth-child(2) {
+  opacity: 0;
+}
+.hamburger-open .bar:nth-child(3) {
+  transform: rotate(-225deg) translateY(7px);
+}
+
+@media only screen and (min-width: 788px) {
+  .hamburger {
+    display: none;
+  }
+}
