+++ static/styles/_navbar.scss
@@ -1,0 +1,120 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+@import url("https://fonts.googleapis.com/css2?family=Kanit:ital,wght@0,400;0,500;1,300;1,400;1,500&display=swap");
+
+:root {
+  --main-color: rgb(214, 238, 254);
+  --secondary-color: rgb(19, 125, 197);
+}
+
+body {
+  font-family: "Kanit", sans-serif;
+}
+
+a {
+  text-decoration: none;
+}
+
+.navbar {
+  height: 100px;
+  min-width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  font-size: 1.5rem;
+  background-color: var(--main-color);
+  position: relative;
+  background: linear-gradient(
+    to top,
+    var(--main-color) 60%,
+    var(--secondary-color)
+  );
+  position: fixed;
+
+  .logo {
+    height: 100%;
+  }
+  .navbar__menu-desktop {
+    display: none;
+  }
+
+  .navbar__menu-mobile {
+    position: absolute;
+    top: 100px;
+    transform: translateX(-100%);
+    transition: 0.5s all ease-in-out;
+    flex-direction: column;
+    width: 100%;
+    background-color: var(--main-color);
+  }
+  .menu__item {
+    list-style: none;
+    height: 100%;
+    width: 100%;
+    transition: 0.2s all ease-in-out;
+    position: relative;
+    // margin-right: 2rem;
+
+    &::before {
+      content: "";
+      position: absolute;
+      width: 0;
+      left: 50%;
+      bottom: 0;
+      height: 3px;
+      background-color: var(--secondary-color);
+      transform: translate(-50%, 100%);
+      transition: 0.2s all ease-in-out;
+    }
+    &:hover::before {
+      width: 100%;
+    }
+    .link {
+      align-items: center;
+      color: var(--secondary-color);
+      width: 100%;
+      height: 100%;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      padding: 1rem;
+      transition: 0.2s all ease-in-out;
+      &:hover {
+        background-color: rgb(19, 125, 197);
+        opacity: 0.5;
+        color: white;
+      }
+    }
+  }
+  .navbar__menu-mobile-open {
+    transform: translateX(0%);
+  }
+}
+
+@media only screen and (min-width: 788px) {
+  .navbar {
+    .navbar__menu-mobile {
+      display: none;
+    }
+    .navbar__menu-desktop {
+      width: 100%;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      gap: 10px;
+      flex-direction: row;
+      height: 100%;
+      .menu__item {
+        height: 100%;
+        width: auto;
+        .link {
+          height: 100%;
+        }
+      }
+    }
+  }
+}
