+++ static/styles/main.scss
@@ -1,0 +1,13 @@
+@import "navbar";
+@import "hamburger";
+@import "header";
+
+button {
+  font-family: "Kanit", sans-serif;
+  background: none;
+  border: 2px solid white;
+  padding: 5px 10px;
+  border-radius: 0.8rem;
+  color: white;
+  font-size: 1.2rem;
+}
