--- styles/_header.scss
@@ -1,49 +1,0 @@
-.header {
-  background: linear-gradient(
-      to bottom,
-      rgba(119, 186, 215, 0.578),
-      rgba(4, 58, 100, 0.578)
-    ),
-    url(../source/images/header-mobile-bg.jpg);
-  height: 110vh;
-  background-position: center;
-  background-size: cover;
-  background-repeat: no-repeat;
-  text-align: center;
-  padding-top: 1px;
-
-  .main-title {
-    color: white;
-    margin-top: 150px;
-    letter-spacing: 5px;
-  }
-  button {
-    position: absolute;
-    bottom: 2rem;
-    transform: translate(-50%, 0);
-    transition: all 0.2s ease-in-out;
-    &:hover {
-      background-color: rgb(255, 255, 255);
-      transform: translate(-50%, -2px);
-      box-shadow: 0 5px 10px -5px rgb(103, 102, 102);
-      padding: 5px 20px;
-      color: var(--secondary-color);
-      border: 5px solid var(--secondary-color);
-    }
-  }
-}
-
-@media only screen and (min-width: 788px) {
-  .header {
-    background: linear-gradient(
-        to bottom,
-        var(--main-color),
-        var(--secondary-color)
-      ),
-      url(../source/images/header-desktop-bg.jpg);
-    height: 80vh;
-    background-position: center;
-    background-size: cover;
-    background-repeat: no-repeat;
-  }
-}
